refactor: submit board phase confirmation dialog
